VACMS-8499 update comments, move drupalAttr to storybook folder.
